// app/types/team.d.ts
export interface TeamMember {
    name: string;
    role: string;
    bio: string;
    image: string;
    year: string;
    contact: string;
    specialties: string[];
    github: string;
    linkedin: string;
  }
  
export interface Milestone {
    year: string;
    title: string;
    description: string;
    
  }